#2312 - Add support for new external recommender v2
- Move v1 recommender to sub-package
