+++ inception/inception-imls-external/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/external/v1/config/ExternalRecommenderPropertiesImpl.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.recommendation.imls.external.config;
+package de.tudarmstadt.ukp.inception.recommendation.imls.external.v1.config;
 
 import static java.time.temporal.ChronoUnit.SECONDS;
