+++ inception/inception-imls-external/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/external/v1/messages/TrainingRequest.java
@@ -15,14 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.recommendation.imls.external.messages;
+package de.tudarmstadt.ukp.inception.recommendation.imls.external.v1.messages;
 
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import de.tudarmstadt.ukp.inception.recommendation.imls.external.model.Document;
-import de.tudarmstadt.ukp.inception.recommendation.imls.external.model.Metadata;
+import de.tudarmstadt.ukp.inception.recommendation.imls.external.v1.model.Document;
+import de.tudarmstadt.ukp.inception.recommendation.imls.external.v1.model.Metadata;
 
 public class TrainingRequest
 {
